Misc fixes noticed by "sparse" or by running gcc with additional 	-W flags
	* libsoup/soup-auth-manager-ntlm.c (ntlm_authorize_post): fix a
	potentially uninitialized variable. (Grumble. gcc needs
	-Wdo-optimization-so-you-can-generate-code-flow-related-warnings-
	but-then-emit-unoptimized-code-for-ease-of-debugging)

	* libsoup/soup-gnutls.c (soup_gnutls_channel_funcs): make this
	static

	* libsoup/soup-uri.c (uri_decoded_copy, uri_normalized_copy): add
	"static". (This doesn't change the generated code; the prototype
	was already declared static and so gcc was treating the function
	as static even though the main declaration *wasn't* declared
	static. I'm not sure if this is a bug in gcc or an oddity of the
	spec, but it's confusing, so...)

	* libsoup/soup-xmlrpc.c (soup_xmlrpc_build_method_response):
	s/FALSE/NULL/

	* libsoup/soup-xmlrpc.h: add G_GNUC_PRINTF to
	soup_xmlrpc_build_format

	* tests/*.c: misc minor fixes, mostly involving missing "const"s
	and "static"s to get better warnings, and then remove some unused
	variables.

	* tests/continue-test.c (do_message): fix a crash when the test
	fails

	* tests/test-utils.h (debug_printf): add G_GNUC_PRINTF to
	prototype

svn path=/trunk/; revision=1123

+++ tests/chunk-test.c
@@ -138,7 +138,7 @@ do_request_test (SoupSession *session, SoupURI *base_uri)
 	}
 	if (msg->request_body->length != length || length != ptd.nwrote) {
 		debug_printf (1, "  sent length mismatch: %d vs %d vs %d\n",
-			      msg->request_body->length, length, ptd.nwrote);
+			      (int)msg->request_body->length, length, ptd.nwrote);
 		errors++;
 	}
